+++ src/main/java/chess/domain/piece/Empty.java
@@ -1,6 +1,7 @@
 package chess.domain.piece;
 
 import chess.domain.piece.coordinate.Coordinate;
+import chess.view.SymbolMatcher;
 
 public class Empty extends Piece {
     public Empty(Team team, Coordinate coordinate) {
