+++ src/main/java/chess/domain/piece/King.java
@@ -1,6 +1,7 @@
 package chess.domain.piece;
 
 import chess.domain.piece.coordinate.Coordinate;
+import chess.view.SymbolMatcher;
 
 public class King extends Piece {
     private static final int MAX_DIFFERENCE_OF_KING = 1;
