+++ src/main/java/chess/domain/piece/Pawn.java
@@ -2,6 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.piece.coordinate.Coordinate;
 
+import chess.view.SymbolMatcher;
 import java.util.List;
 
 public class Pawn extends Piece {
