+++ src/main/java/chess/domain/piece/Piece.java
@@ -2,6 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.piece.coordinate.Coordinate;
 
+import chess.view.SymbolMatcher;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
