+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,6 +1,7 @@
 package chess.domain.piece;
 
 import chess.domain.piece.coordinate.Coordinate;
+import chess.view.SymbolMatcher;
 
 public class Queen extends Piece {
     public Queen(Team team, Coordinate coordinate) {
