+++ src/test/java/chess/piece/EmptyTest.java
@@ -2,7 +2,7 @@ package chess.piece;
 
 import chess.domain.piece.Empty;
 import chess.domain.piece.Piece;
-import chess.domain.piece.SymbolMatcher;
+import chess.view.SymbolMatcher;
 import chess.domain.piece.Team;
 import chess.domain.piece.coordinate.Coordinate;
 import org.junit.jupiter.api.DisplayNameGeneration;
