+++ src/test/java/chess/piece/PieceMatcherTest.java
@@ -9,7 +9,7 @@ import chess.domain.piece.Piece;
 import chess.domain.piece.PieceMatcher;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
-import chess.domain.piece.SymbolMatcher;
+import chess.view.SymbolMatcher;
 import chess.domain.piece.Team;
 import chess.domain.piece.coordinate.Coordinate;
 import org.junit.jupiter.api.BeforeEach;
